Terminal: graphical settings for configuring themes.
Implements #16835: Make colour schemes modular.
Implements #15936: Terminal settings should allow to drop a colour.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/5950
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/apps/terminal/PrefWindow.h
@@ -14,10 +14,10 @@
 #include <Window.h>
 
 // local messages
-const ulong MSG_DEFAULTS_PRESSED = 'defl';
-const ulong MSG_SAVEAS_PRESSED = 'canl';
-const ulong MSG_REVERT_PRESSED = 'revt';
-const ulong MSG_PREF_CLOSED = 'mspc';
+const uint32 MSG_DEFAULTS_PRESSED = 'defl';
+const uint32 MSG_SAVEAS_PRESSED = 'canl';
+const uint32 MSG_REVERT_PRESSED = 'revt';
+const uint32 MSG_PREF_CLOSED = 'mspc';
 
 class BButton;
 class BFilePanel;
